requirements.txt
Flask==2.2.3
pymongo[srv]==4.1.1
bcrypt==4.0.1
PyJWT==2.8.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.0
gunicorn==23.0.0
eventlet==0.33.1
werkzeug==2.2.2